Check to make sure qbar is the converged value
This commit adjust the logic to only plot those values of qbar that are
negative and converged (the last iter of the problem for a given group).
This is done by checking against the source components to look for those
that have higher iteration values.  If no source components have higher
iteration values, then the negative qbar is the converged value.

Added functions getlabelgrp and getlabelitr to get the components of
interest from a VTK table entry output with the MPACT "vtk_iter_edits"
branch.