Changed masses from hidden MCNP values to NIST
NIST values were captured from the website and a converter script was
created (Convert_NIST_Data.py) to convert the linearized website layout
to a dictionary setup appropriate for use with Python.

As such, the NuclearData library was updated and the README.md file was
updated to contain unredacted output.